+++ app/renderer/constants/routes.js
@@ -1,6 +1,7 @@
 // @flow
 export default {
   ROOT: '/',
+  WAIT: '/wait',
   SETTINGS: {
     ROOT: '/settings',
     GENERAL: '/settings/general',