+++ app/renderer/pages/devPage/index.js
@@ -37,9 +37,10 @@ const DevPage = ({
   nodeSystem,
   syncStream,
   syncRemoteStream,
+  networkStatusStore,
 }) => {
   const { chain, name, version, health } = nodeSystem;
-  const networkStatus = `${chain} ${version} (status:${health.message}, sync:${
+  const networkStatusLabel = `${chain} ${version} (status:${health.message}, sync:${
     health.isSyncing
   }, peers:${health.peers}, name:${name})`;
 
@@ -60,7 +61,7 @@ const DevPage = ({
         </PageTitle>
         <Flex>
           <Button onClick={onNetworkStatusClick}>Get Network Status</Button>
-          <div>{networkStatus}</div>
+          <div>{networkStatusLabel}</div>
         </Flex>
         <Flex>
           <Button onClick={onGetHeaderClick}>Get Header</Button>
@@ -72,6 +73,7 @@ const DevPage = ({
         </Flex>
         <Flex>
           <Button onClick={onRestartNodeClick}>Restart node</Button>
+          <div>{JSON.stringify(networkStatusStore)}</div>
         </Flex>
         <Flex>
           <Button onClick={onStreamStart}>Start stream</Button>