+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Improvements
 
+- able to detect app state and able to show spinner on exit
+
 # v0.3.0 (2019.01.27)
 
 ### Features